Make probability with which an action is executed configurable
The probabilities with which each action was executed in each test cycle
was hard coded. This made it hard to create different test scenarios.

Now the probabilities can be defined in the config file. By setting them
to 0 or 1, they can be configured to never or to always run,
respectively. Note, that some actions depend on other actions to create
the state required for themselves.

Change-Id: I23b436cda01706b71fa309d4309f133b9181cc92