Parse config recursively
So far the config was only parsed two levels deep, not allowing to apply
defaults deeper than that. This would lead to issues as soon as there is
for example a second parameter for each action that can be configured.

Now the parser goes through the config recursively allowing an
indefinite depth of the config dictionary.

Change-Id: I7a0eed1e0266dbe52b19716df81e93881dceebf7