Allow to configure a list of projects known from the start
An instance ('simulated user') is only using projects that it knows of.
The list of known projects could be grown by either creating a project
or by querying projects, in which case one of the returned projects in
the query was added to the list of known projects.

However, for some usecases certain projects should be tested against,
e.g. they have a specific property like unusually high number of refs.
Such a directed test was not possible yet.

This change allows to configure a list of projects that each simulated
user knows from the beginning of the test. If the querying and creation
of projects is deactivated (by setting the respective probability to 0),
the test will only work with these projects.

Change-Id: I4f9e58c9b12930251aad73dcc7962a534541a036