Move actions done before a test run out of the constructor
The initial delay and cloning of projects were done, when a test client
object was instantiated. This didn't seem like the proper place to do it
and also led to a large constructor.

Now these steps are performed in the prerun- method, which is called before
the run()-method is called.

Change-Id: Ic6b3766f4ee12e3c0d4d4633c97fb0c7177b143c